refactor: Validator를 views폴더에 이동
