+++ __tests__/Application.js
@@ -1,8 +1,8 @@
 import App from '../src/App.js';
 import { mockQuestions, mockRandoms } from '../src/helpers/testUtils.js';
 
-describe('애플리케이션 테스트', () => {
-  test('우승자가 한명일때', async () => {
+describe('애플리케이션', () => {
+  test('우승자가 한명인 경우', async () => {
     const inputs = ['a,b,c', '1'];
     const randomNumbers = [1, 1, 9];
